+++ web/src/components/Tables/PolicyTable/PolicyTable.js
@@ -1,7 +1,7 @@
 import React from "react";
 import {DownOutlined, EditOutlined, MinusOutlined, UpOutlined} from '@ant-design/icons';
 import {Button, Input, Row, Select, Table, Tooltip} from 'antd';
-import * as Setting from "./Setting";
+import * as Setting from "../../../utils/Setting";
 
 const { Option } = Select;
