+++ web/src/containers/ModelPage/ModelPage.js
@@ -1,10 +1,10 @@
 import React from "react";
-import * as Setting from "./Setting";
-import * as Backend from "./Backend";
+import * as Setting from "../../utils/Setting";
+import * as Backend from "../../utils/Backend";
 import {Button, Card, Col, Input, Row, Select, Tag} from "antd";
 import {Controlled as CodeMirror} from 'react-codemirror2'
 import "codemirror/lib/codemirror.css"
-import ModelEditor from "./ModelEditor";
+import ModelEditor from "../ModelEditor/ModelEditor";
 
 require("codemirror/mode/properties/properties");
 
@@ -27,11 +27,12 @@ class ModelPage extends React.Component {
   getModel() {
     Backend.getModel(this.state.modelId)
       .then((res) => {
-        let model = res;
-        model.text = this.parseModelText(model.text);
-        this.setState({
-          model: model,
-        });
+        console.log("aaa",res);
+        // let model = res;
+        // model.text = this.parseModelText(model.text);
+        // this.setState({
+        //   model: model,
+        // });
         }
       );
   }
